Simplify BuildResult enum
This removes the BuildResult::Failure variant. It was used
interchangeably with BuildResult::Success, and only returned in rustc()
routine when executed in-process compiler panicked or returned with a
non-zero exit code.

Since internally the compiler driver also panics on compile failure, the
BuildResult::{Failure, Success} variant distinction didn't give any
meaningful information, as the driver treats ICEs and regular compile
failures the same way. In both cases BuildResult::Failure was returned.

More correct solution would be not to panic on regular compile failure
in run_compiler, so the caller can either catch a panic and deduce that
an ICE was encountered, or receive a non-zero (101) exit code in case
errors were encountered.
Then a BuildResult::Err variant would be returned on an ICE, or
BuildResult::Success otherwise.